fixed clippy warnings except dead-code
