+++ src/sphere.rs
@@ -1,4 +1,3 @@
-use crate::distance::Euclidean;
 use num_traits::Float;
 use num_traits::FromPrimitive;
 use num_traits::Zero;
@@ -6,6 +5,8 @@ use std::ops::AddAssign;
 use std::ops::DivAssign;
 use std::ops::MulAssign;
 use std::ops::SubAssign;
+
+use crate::distance::Euclidean;
 
 #[derive(Debug)]
 pub struct Sphere<T> {
@@ -18,11 +19,16 @@ where
     T: Float + Zero + FromPrimitive + AddAssign + SubAssign + DivAssign + MulAssign,
 {
     pub fn new(center: Vec<T>, radius: T) -> Sphere<T> {
-        Sphere { center: center, radius }
+        Sphere { center, radius }
     }
 
-    pub fn from_point(point: &Vec<T>) -> Sphere<T> {
-        Sphere::new(point.clone(), T::zero())
+    pub fn from_point(point: &[T]) -> Sphere<T> {
+        Sphere::new(point.to_owned(), T::zero())
+    }
+
+    pub fn intersects(&self, sphere: &Sphere<T>) -> bool {
+        let distance = Euclidean::distance(&self.center, &sphere.center);
+        distance - (self.radius + sphere.radius) <= T::zero()
     }
 }
 
@@ -34,5 +40,33 @@ mod tests {
     pub fn test_from_point(){
         let sphere = Sphere::from_point(&vec![1.,1.]);
         assert_eq!(sphere.radius, 0.);
+    }
+
+    #[test]
+    pub fn test_sphere_intersects_sphere() {
+        let sphere1 = Sphere::new(vec![0., 0.], 10.);
+        let sphere2 = Sphere::new(vec![15., 15.], 15.);
+        assert!(sphere1.intersects(&sphere2));
+    }
+
+    #[test]
+    pub fn test_sphere_doesnot_intersect_sphere() {
+        let sphere1 = Sphere::new(vec![450., 150.], 50.);
+        let sphere2 = Sphere::new(vec![530., 220.], 50.);
+        assert_eq!(sphere1.intersects(&sphere2), false);
+    }
+
+    #[test]
+    pub fn test_sphere_intersects_its_clone() {
+        let sphere1 = Sphere::new(vec![0., 0.], 10.);
+        let sphere2 = Sphere::new(vec![0., 0.], 10.);
+        assert!(sphere1.intersects(&sphere2));
+    }
+
+    #[test]
+    pub fn test_sphere_intersects_smaller_sphere() {
+        let sphere1 = Sphere::new(vec![0., 0.], 10.);
+        let sphere2 = Sphere::new(vec![10., 10.], 100.);
+        assert!(sphere1.intersects(&sphere2));
     }
 }
